examples/tynacorp/pyproject.toml
[tool.poetry]
name = "tynacorp"
version = "0.1.0"
package-mode = false
description = "An example fastrpc server app/library"
authors = [
    "Andrew Stefanich <andrewstefanich@gmail.com>",
]

[tool.poetry.dependencies]
python = "^3.12"
bpython = "^0.24"
fastapi-pagination = "^0.12.24"
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
gunicorn = "^22.0.0"
returns = "^0.23.0"
typer = "^0.12.3"
aiofiles = "^23.2.1"
fastapi = "^0.111.1"
pydantic = "^2.9.2"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
debugpy = "^1.8.1"
pre-commit = "^3.7.0"
pytest = "^8.3.1"
pytest-asyncio = "^0.23.8"
pytest-parametrization = "^2022.2.1"
ruff = "^0.6.6"

[tool.ruff]
lint.ignore = ["E731"] # allow assignment of lambdas to a symbol

[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api"
